+++ src/downpodia/config.nim
@@ -1,0 +1,1 @@
+const userAgent* = "Mozilla/5.0 (X11; Linux x86_64; rv\":7\"8.0) Gecko/20100101 Firefox/78.0"